requirements.txt
click==7.1.2
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
